se agrego fetch y librerias

+++ js/carrito.js
@@ -69,6 +69,23 @@ function actualizarBotonesEliminar() {
 }
 
 function eliminarDelCarrito(e) {
+    Toastify({
+        text: "Producto eliminado",
+        duration: 3000,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background: "linear-gradient(to right, #03174d, #3b6fff)",
+          borderRadius: "2rem",
+        },
+        offset: {
+            x: '1.5rem', // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+            y: '2rem' // vertical axis - can be a number or a string indicating unity. eg: '2em'
+          },
+        onClick: function(){} // Callback after click
+      }).showToast();
     const idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
 
@@ -79,9 +96,23 @@ function eliminarDelCarrito(e) {
 
 botonVaciar.addEventListener("click", vaciarCarrito);
 function vaciarCarrito() {
-    productosEnCarrito.length = 0;
-    cargarProductosCarrito();
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+
+    Swal.fire({
+        title: 'Estas seguro',
+        icon: 'question',
+        html: 'Se van a borrar todos tus productos.',
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: 'Sí',
+        cancelButtonText: 'No',
+      }).then((result) => {
+
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            cargarProductosCarrito();
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+      }
+    })
 }
 
 function actualizarTotal() {
